+++ ocr_wrapper/azure.py
@@ -33,36 +33,65 @@ def requires_azure(fn):
     return wrapper_decocator
 
 
+def _discretize_angle_to_90_deg(rotation: float) -> int:
+    """Discretize an angle to the nearest 90 degrees"""
+    return int(((rotation + 45) // 90 * 90) % 360)
+
+
+def _determine_endpoint_and_key(endpoint: Optional[str], key: Optional[str]) -> tuple[str, str]:
+    """Determine the endpoint and key to be used.
+
+    If endpoint and key are both None, the values are looked up in the environment variables AZURE_OCR_ENDPOINT and
+    AZURE_OCR_KEY. If these are not set, the values are read from the file ~/.config/azure/ocr_credentials.json.
+
+    If only one of endpoint and key is None, only the other one is looked up in the environment variables and the file.
+    """
+    if endpoint is None:
+        endpoint = os.environ.get("AZURE_OCR_ENDPOINT")
+    if key is None:
+        key = os.environ.get("AZURE_OCR_KEY")
+    if endpoint is None or key is None:
+        if os.path.exists(os.path.expanduser("~/.config/azure/ocr_credentials.json")):
+            with open(os.path.expanduser("~/.config/azure/ocr_credentials.json")) as f:
+                data = json.load(f)
+                endpoint = endpoint or data.get("endpoint")
+                key = key or data.get("key")
+    if endpoint is None or key is None:
+        raise Exception("Azure endpoint and key must be specified via some means")
+
+    return endpoint, key
+
+
 class AzureOCR(OcrWrapper):
     @requires_azure
     def __init__(
         self,
         *,
         cache_file: Optional[str] = None,
-        max_size: Optional[int] = 1024,
+        max_size: Optional[int] = None,
+        auto_rotate: bool = False,
         ocr_samples: int = 1,
+        endpoint: Optional[str] = None,
+        key: Optional[str] = None,
         verbose: bool = False,
     ):
         super().__init__(
             cache_file=cache_file,
             max_size=max_size,
+            auto_rotate=auto_rotate,
             ocr_samples=ocr_samples,
+            supports_multi_samples=False,
             verbose=verbose,
         )
-        keyfile = "~/.config/azure/ocr_credentials.json"
-        with open(os.path.expanduser(keyfile), mode="r") as f:
-            ocr_credentials = json.load(f)
-        self.client = ComputerVisionClient(
-            endpoint=ocr_credentials["endpoint"],
-            credentials=CognitiveServicesCredentials(ocr_credentials["key1"]),
-        )
+        endpoint, key = _determine_endpoint_and_key(endpoint, key)
+        self.client = ComputerVisionClient(endpoint, CognitiveServicesCredentials(key))
 
     @requires_azure
     def _get_ocr_response(self, img: Image.Image):
         """Gets the OCR response from the Azure. Uses cached response if a cache file has been specified and the
         document has been OCRed already"""
         # Pack image in correct format
-        img_bytes = self._pil_img_to_compressed(img)
+        img_bytes = self._pil_img_to_compressed(img, compression="png")
         img_stream = BytesIO(img_bytes)
 
         # Try to get cached response
@@ -92,5 +121,10 @@ class AzureOCR(OcrWrapper):
             for line in annotation.lines:
                 for word in line.words:
                     bbox = BBox.from_pixels(word.bounding_box, original_size=img.size)
-                    result.append({"bbox": bbox, "text": word.text})
+                    result.append({"bbox": bbox, "text": word.text, "confidence": word.confidence})
+
+        # Determine rotation of document
+        page_rotation = response.analyze_result.read_results[0].angle
+        self.extra["document_rotation"] = _discretize_angle_to_90_deg(page_rotation)
+
         return result