+++ ocr_wrapper/__init__.py
@@ -1,10 +1,13 @@
 from .aws import AwsOCR
 from .azure import AzureOCR
-from .google_ocr import GoogleOCR
-from .paddleocr import PaddleOCR
-from .easy_ocr import EasyOCR
 from .bbox import BBox, draw_bboxes, get_label2color_dict
+from .easy_ocr import EasyOCR
+from .google_ocr import GoogleOCR
 from .ocr_wrapper import OcrWrapper
+from .paddleocr import PaddleOCR
+
+# Important as last import, because it depends on the other modules
+from .autoselect import autoselect_ocr_engine  # isort:skip
 
 __all__ = [
     "AwsOCR",
@@ -15,5 +18,6 @@ __all__ = [
     "BBox",
     "draw_bboxes",
     "get_label2color_dict",
+    "autoselect_ocr_engine",
     "OcrWrapper",
 ]