+++ tests/test_googleocr.py
@@ -1,7 +1,8 @@
 import os
-from PIL import Image
-from ocr_wrapper import GoogleOCR
+
 import pytest
+from ocr_wrapper import GoogleOCR
+from PIL import Image
 
 filedir = os.path.dirname(__file__)
 DATA_DIR = os.path.join(filedir, "data")
@@ -60,6 +61,18 @@ def test_google_orc_single_sample():
     text = " ".join([str(r.text) for r in res])
     assert text == "This is a test ."
     assert len(extra["confidences"][0]) == len(res)
+
+
+@pytest.mark.parametrize("rotation_angle", [0.5, -2.2, 3.5, -9.5])
+@pytest.mark.parametrize("ocr_system", ["ocr", "ocr_with_auto_rotate"])
+def test_tilt_correction(rotation_angle, ocr_system, request):
+    ocr = request.getfixturevalue(
+        ocr_system
+    )  # Get the fixture by name (has to be done this way because the fixture is parametrized)
+    with Image.open(os.path.join(DATA_DIR, "ocr_samples.png")) as img:
+        rot_img = img.rotate(rotation_angle, expand=True, fillcolor="white")
+        _, extra = ocr.ocr(rot_img, return_extra=True)
+        assert extra["tilt_angle"] == pytest.approx(rotation_angle, abs=0.01)
 
 
 @pytest.mark.parametrize("filename, rotation", rotation_test_documents)