+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The version numbers are according to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added the output of confidence scores to GoogleOCR
 - Added multiple OCR passes to improve OCR reliability
-- Adds an environment variable `OCR_WRAPPER_CACHE_FILE` to specify an ocr cache file globally
-
+- Added an environment variable `OCR_WRAPPER_CACHE_FILE` to specify an ocr cache file globally
+- Added an `autoselect_ocr_engine` function that selects the correct engine depending on the `OCR_PROVIDER` environment variable
 ### Changed
 - Changed GoogleOCR to use WebP instead of PNG to transfer images to the cloud (reduces amount of transferred data by ~ 1/2)
 ### Fixed